+++ core-ng/src/main/java/core/framework/log/message/EventMessage.java
@@ -11,12 +11,12 @@ import java.util.Map;
 public class EventMessage {
     @Property(name = "id")
     public String id;
-    @Property(name = "timestamp")
-    public Instant timestamp;
+    @Property(name = "date")
+    public Instant date;
     @Property(name = "app")
     public String app;
-    @Property(name = "event_time")
-    public Instant eventTime;
+    @Property(name = "received_time")
+    public Instant receivedTime;
     @Property(name = "result")
     public String result;
     @Property(name = "action")