+++ log-collector/src/main/java/core/log/web/SendEventRequestValidator.java
@@ -7,14 +7,14 @@ import static core.framework.util.Strings.format;
 /**
  * @author neo
  */
-public class CollectEventRequestValidator {
+public class SendEventRequestValidator {
     private static final int MAX_CONTEXT_VALUE_LENGTH = 1000;
 
-    void validate(CollectEventRequest request) {
-        for (CollectEventRequest.Event event : request.events) {
-            if (event.result == CollectEventRequest.Result.OK && event.action == null)
+    void validate(SendEventRequest request) {
+        for (SendEventRequest.Event event : request.events) {
+            if (event.result == SendEventRequest.Result.OK && event.action == null)
                 throw new BadRequestException("action must not be null if result is OK");
-            if ((event.result == CollectEventRequest.Result.WARN || event.result == CollectEventRequest.Result.ERROR)
+            if ((event.result == SendEventRequest.Result.WARN || event.result == SendEventRequest.Result.ERROR)
                     && event.errorCode == null)
                 throw new BadRequestException("errorCode must not be null if result is WARN/ERROR");